model_handler_template/requirements.txt
openai
grpcio
grpcio-tools
protobuf
pydantic
langchain
langchain-openai
pymongo
datetime
pytz